+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Unreleased changes
+- Added method `getFinalizedBlocksFrom` for retrieving finalized blocks starting from a given height.
+- Added method `getBlocksAtHeight` for retrieving live blocks at a given height.
 - Added method `findAtLowestHeight` for finding the earliest block satisfying some condition.
 - Added method `findAccountCreation` for finding the block in which an account was created.
 - Make the `energy` parameter for invoking an instance `Optional`.